Address review: move password_policy_violation to util, fix docstrings
Move password_policy_violation out of const.py (which is for constants, not logic) into
util/__init__.py, keeping the policy constants in const. Update the two call sites. Also correct
the add_user / set_password docstrings to state the 12-128 character range.
